update: handleLogout location changed
